+++ include/nfx/Graphics.h
@@ -32,6 +32,7 @@
 #include "nfx/graphics/gl/pipeline/passes/PostProcessPass.h"
 #include "nfx/graphics/gl/pipeline/passes/PresentPass.h"
 #include "nfx/graphics/gl/pipeline/passes/RenderPass.h"
+#include "nfx/graphics/gl/pipeline/passes/SkyboxPass.h"
 #include "nfx/graphics/gl/pipeline/passes/SpotShadowPass.h"
 #include "nfx/graphics/gl/pipeline/passes/TransparentPass.h"
 #include "nfx/graphics/gl/pipeline/queue/RenderCommand.h"